refactor: rename libs folder to src
